Add first working version of PMD driver.
The driver is implemented as a nodelet (similarly to OpenNI driver
in ROS). It publishes only depth images and camera calibration
information as received from the SDK driver.

Included is a launch file (again, inspired by openni_launch package in
ROS), which brings up a nodelet manager with PMD driver, image rectifier
and point cloud generator.

The resulting point clouds seem to be wrong, this issue needs to be explored.